+++ base_modelo.py
@@ -1,0 +1,101 @@
+import torch
+import torch.nn as nn
+import torch.optim as optim
+import numpy as np
+import json
+from nltk.stem.porter import PorterStemmer
+import nltk
+
+nltk.download('punkt')
+
+# Funciones de preprocesamiento
+def tokenize(sentence):
+    return nltk.word_tokenize(sentence)
+
+def stem(word):
+    return stemmer.stem(word.lower())
+
+def bag_of_words(tokenized_sentence, words):
+    sentence_words = [stem(word) for word in tokenized_sentence]
+    bag = np.zeros(len(words), dtype=np.float32)
+    for idx, w in enumerate(words):
+        if w in sentence_words:
+            bag[idx] = 1
+    return bag
+
+# Cargar datos de intents
+with open('intents.json', 'r', encoding='utf-8') as f:
+    intents = json.load(f)
+
+stemmer = PorterStemmer()
+
+all_words = []
+tags = []
+xy = []
+
+for intent in intents['intents']:
+    tag = intent['tag']
+    tags.append(tag)
+    for pattern in intent['patterns']:
+        w = tokenize(pattern)
+        all_words.extend(w)
+        xy.append((w, tag))
+
+ignore_words = ['?', '!', '.', ',']
+all_words = [stem(w) for w in all_words if w not in ignore_words]
+all_words = sorted(set(all_words))
+tags = sorted(set(tags))
+
+X_train = []
+y_train = []
+
+for (pattern_sentence, tag) in xy:
+    bag = bag_of_words(pattern_sentence, all_words)
+    X_train.append(bag)
+    label = tags.index(tag)
+    y_train.append(label)
+
+X_train = np.array(X_train)
+y_train = np.array(y_train)
+
+# Definir el modelo
+class NeuralNet(nn.Module):
+    def __init__(self, input_size, hidden_size, output_size):
+        super(NeuralNet, self).__init__()
+        self.l1 = nn.Linear(input_size, hidden_size)
+        self.l2 = nn.Linear(hidden_size, hidden_size)
+        self.l3 = nn.Linear(hidden_size, output_size)
+        self.relu = nn.ReLU()
+    
+    def forward(self, x):
+        out = self.l1(x)
+        out = self.relu(out)
+        out = self.l2(out)
+        out = self.relu(out)
+        out = self.l3(out)
+        return out
+
+input_size = len(all_words)
+hidden_size = 8
+output_size = len(tags)
+learning_rate = 0.001
+num_epochs = 1000
+
+model = NeuralNet(input_size, hidden_size, output_size)
+
+criterion = nn.CrossEntropyLoss()
+optimizer = optim.Adam(model.parameters(), lr=learning_rate)
+
+for epoch in range(num_epochs):
+    outputs = model(torch.from_numpy(X_train).float())
+    loss = criterion(outputs, torch.from_numpy(y_train).long())
+    
+    optimizer.zero_grad()
+    loss.backward()
+    optimizer.step()
+    
+    if (epoch + 1) % 100 == 0:
+        print(f'Epoch [{epoch + 1}/{num_epochs}], Loss: {loss.item():.4f}')
+
+# Guardar el modelo
+torch.save(model.state_dict(), 'model.pth')
