+++ chatbot.py
@@ -1,111 +1,22 @@
 import torch
-import torch.nn as nn
-import torch.optim as optim
-import numpy as np
 import json
 import random
 import matplotlib.pyplot as plt
 import datetime
+from base_modelo import tokenize, stem, bag_of_words, NeuralNet
 
-
-with open('intents.json', 'r', encoding='utf-8') as f:
-    intents = json.load(f)
-
-
-import nltk
-from nltk.stem.porter import PorterStemmer
-nltk.download('punkt')
-stemmer = PorterStemmer()
-
-def tokenize(sentence):
-    return nltk.word_tokenize(sentence)
-
-def stem(word):
-    return stemmer.stem(word.lower())
-
-def bag_of_words(tokenized_sentence, words):
-    sentence_words = [stem(word) for word in tokenized_sentence]
-    bag = np.zeros(len(words), dtype=np.float32)
-    for idx, w in enumerate(words):
-        if w in sentence_words:
-            bag[idx] = 1
-    return bag
-
-
-all_words = []
-tags = []
-xy = []
-
-for intent in intents['intents']:
-    tag = intent['tag']
-    tags.append(tag)
-    for pattern in intent['patterns']:
-        w = tokenize(pattern)
-        all_words.extend(w)
-        xy.append((w, tag))
-
-ignore_words = ['?', '!', '.', ',']
-all_words = [stem(w) for w in all_words if w not in ignore_words]
-all_words = sorted(set(all_words))
-tags = sorted(set(tags))
-
-X_train = []
-y_train = []
-
-for (pattern_sentence, tag) in xy:
-    bag = bag_of_words(pattern_sentence, all_words)
-    X_train.append(bag)
-    label = tags.index(tag)
-    y_train.append(label)
-
-X_train = np.array(X_train)
-y_train = np.array(y_train)
-
-
-class NeuralNet(nn.Module):
-    def __init__(self, input_size, hidden_size, output_size):
-        super(NeuralNet, self).__init__()
-        self.l1 = nn.Linear(input_size, hidden_size)
-        self.l2 = nn.Linear(hidden_size, hidden_size)
-        self.l3 = nn.Linear(hidden_size, output_size)
-        self.relu = nn.ReLU()
-    
-    def forward(self, x):
-        out = self.l1(x)
-        out = self.relu(out)
-        out = self.l2(out)
-        out = self.relu(out)
-        out = self.l3(out)
-        return out
-
-
-input_size = len(all_words)
-hidden_size = 8
-output_size = len(tags)
-learning_rate = 0.001
-num_epochs = 1000
-
-model = NeuralNet(input_size, hidden_size, output_size)
-
-
-criterion = nn.CrossEntropyLoss()
-optimizer = optim.Adam(model.parameters(), lr=learning_rate)
-
-
-for epoch in range(num_epochs):
-    outputs = model(torch.from_numpy(X_train).float())
-    loss = criterion(outputs, torch.from_numpy(y_train).long())
-    
-    optimizer.zero_grad()
-    loss.backward()
-    optimizer.step()
-    
-    if (epoch + 1) % 100 == 0:
-        print(f'Epoch [{epoch + 1}/{num_epochs}], Loss: {loss.item():.4f}')
-
-
+def load_model():
+    # Cargar el modelo entrenado desde el archivo 'model.pth'
+    input_size = len(all_words)
+    hidden_size = 8
+    output_size = len(tags)
+    model = NeuralNet(input_size, hidden_size, output_size)
+    model.load_state_dict(torch.load('model.pth'))
+    model.eval()
+    return model
 
 def predict_class(sentence):
+    # Predecir la clase (intento) para una oración dada
     sentence = tokenize(sentence)
     X = bag_of_words(sentence, all_words)
     X = torch.from_numpy(X).float()
@@ -114,7 +25,9 @@ def predict_class(sentence):
     tag = tags[predicted.item()]
     return tag
 
-def generar_curva_consumo(consumo):
+def generar_curva_consumo():
+    # Generar una curva de consumo aleatoria y guardarla como imagen PNG
+    consumo = [random.randint(5, 30) for _ in range(7)]  # Genera 7 valores aleatorios entre 5 y 30
     plt.figure(figsize=(10, 5))
     plt.plot(consumo, marker='o', linestyle='-', color='b')
     plt.title("Curva de Consumo")
@@ -126,13 +39,23 @@ def generar_curva_consumo(consumo):
     return "Tu curva de consumo ha sido generada. Revisa el archivo 'curva_consumo.png'."
 
 def reportar_corte(ubicacion):
+    # Registrar un reporte de corte en el archivo 'reportes_cortes.txt'
     fecha = datetime.datetime.now()
     reporte = f"Corte reportado en {ubicacion} el {fecha}. Se agendará una visita."
     with open("reportes_cortes.txt", "a") as file:
         file.write(reporte + "\n")
     return reporte
 
+def obtener_curvas_aleatorias(suministro):
+    # Obtener curvas de consumo aleatorias basadas en el suministro dado
+    curvas = {
+        "343223": [1, 2, 3, 4, 5],
+        "85456": [2, 4, 6, 8]
+    }
+    return curvas.get(suministro, [])
+
 def chatbot_response(msg):
+    # Generar una respuesta del chatbot basada en el mensaje del usuario
     tag = predict_class(msg)
     
     for intent in intents['intents']:
@@ -140,16 +63,22 @@ def chatbot_response(msg):
             response = random.choice(intent['responses'])
             
             if tag == "curva_consumo":
-                consumo = [10, 20, 25, 25, 20, 15, 5]  # Simulación de datos de consumo
-                response += "\n" + generar_curva_consumo(consumo)
+                response += "\n" + generar_curva_consumo()
             elif tag == "reportar_corte":
                 response += "\n" + reportar_corte("Ubicación del usuario")
+            elif tag == "reducir_planilla":
+                response += "\n" + "Algunas formas de reducir tu consumo incluyen: \n- Usar electrodomésticos eficientes.\n- Apagar luces que no uses.\n- Aprovechar la luz natural."
+            elif tag == "curvas_aleatorias":
+                suministro = msg.split()[-1]  # Extraer el suministro de la última palabra en el mensaje
+                curvas = obtener_curvas_aleatorias(suministro)
+                response += "\nCurvas para el suministro: " + ", ".join(map(str, curvas))
             
             return response
 
     return "No entendí tu mensaje."
 
 def chat():
+    # Función principal para iniciar el chat con el usuario
     print("Chatbot para empresa eléctrica. Escribe 'salir' para terminar.")
     while True:
         msg = input("Tú: ")
@@ -160,9 +89,22 @@ def chat():
         print("Bot:", response)
 
 if __name__ == "__main__":
+    # Cargar intents y preprocesar datos
+    with open('intents.json', 'r', encoding='utf-8') as f:
+        intents = json.load(f)
+    all_words = []
+    tags = []
+    for intent in intents['intents']:
+        tag = intent['tag']
+        tags.append(tag)
+        for pattern in intent['patterns']:
+            w = tokenize(pattern)
+            all_words.extend(w)
+    ignore_words = ['?', '!', '.', ',']
+    all_words = [stem(w) for w in all_words if w not in ignore_words]
+    all_words = sorted(set(all_words))
+    tags = sorted(set(tags))
+
+    # Cargar el modelo y iniciar el chat
+    model = load_model()
     chat()
-
-
-
-
-
